refactor: import 경로 정리 및 타입 구조 개선
