+++ src/entities/meetings/types/index.ts
