+++ src/features/meetings/model/hook/useCancelJoinMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import meetingKeys from '@/entities/meetings/model/meeting.keys';
-import type { MeetingDetailInfo } from '@/entities/meetings/model/types';
+import type { MeetingDetailInfo } from '@/entities/meetings/types';
 import cancelJoin from '../../api/cancelJoin';
 import type { CancelJoinRequest, CancelJoinResponse, MutationCtx } from '../types';
