+++ src/features/meetings/model/hook/useJoinMeetingMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import meetingKeys from '@/entities/meetings/model/meeting.keys';
-import type { MeetingDetailInfo } from '@/entities/meetings/model/types';
+import type { MeetingDetailInfo } from '@/entities/meetings/types';
 import joinMeeting from '../../api/joinMeeting';
 import type { JoinMeetingRequest, JoinMeetingResponse, MutationCtx } from '../types';
