+++ src/widgets/find-meeting/ui/FindSorterDropdown.tsx
@@ -1,8 +1,7 @@
 import { DropdownMenuContent, DropdownMenuGroup, DropdownMenuItem } from '@/shared';
-import { SortType } from '@/entities/meetings';
+import { SortType } from '@/entities/meetings/types';
+import { useSorterStore, useFilterStore } from '@/entities/meetings';
 import { SorterState } from '../model/types';
-import useSorterStore from '../model/hooks/useSorterStore';
-import useFilterStore from '../model/hooks/useFilterStore';
 
 export default function FindSorterDropdown() {
   const setSelectedSorter = useSorterStore((state) => state.setSelectedSorter);
