+++ src/widgets/home/HomeMeetingSection.tsx
@@ -2,8 +2,8 @@
 
 import { DrinkType } from '@/shared';
 import { useUserStore } from '@/entities/user';
-import { HomeListKind, MeetingCardInfo } from '@/entities/meetings/model/types';
-import MeetingCardCarousel from '@/entities/meetings/ui/MeetingCardCarousel';
+import { HomeListKind, MeetingCardInfo } from '@/entities/meetings/types';
+import { MeetingCardCarousel } from '@/widgets';
 import { useHomeMeetingsQuery } from '@/entities/meetings';
 
 function emptyCopy(kind: HomeListKind) {
